Cleans, builds, and runs with no warnings
