--- search-java-indexer-demo/src/resources/com/microsoft/azure/search/samples/app/config.properties
@@ -1,9 +1,0 @@
-SearchServiceName=[ SEARCH SERVICE NAME]
-SearchServiceApiKey=[SEARCH SERVICE API_KEY]
-IndexName=hotels
-ApiVersion=2019-05-06
-DataSourceName=hotels-datasource
-IndexerName=hotels-indexer
-DataSourceConnectionString={ "connectionString" : 'AccountEndpoint=https://hotels-docbb.documents.azure.com:443/;AccountKey=4UPsNZyFAjgZ1tzHPGZaxS09XcwLrIawbXBWk6IixcxJoSePTcjBn0mi53XiKWu8MaUgowUhIovOv7kjksqAug==;Database=SampleData' }
-DataSourceType=cosmosdb
-DataSourceTable=hotels
